Sem 8/Code/requirements.txt
Flask
pandas
pinecone-client
google-generativeai
langchain
langchain-google-genai
langchain-pinecone
langgraph
markdown
PyPDF2
sumy
fpdf
python-dotenv
